% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateVTT.R
\name{translateVTT}
\alias{translateVTT}
\title{translateVTT}
\usage{
translateVTT(fileName, sourceLang = "en", destLang, apikey)
}
\arguments{
\item{fileName}{Path to the file you want to translate. Only one file name is
accepted at the time}

\item{sourceLang}{Source language code, 'en' is default value.}

\item{destLang}{Destination language for translator function. Only one at the
time}

\item{apikey}{Google Translate API key}
}
\value{
Function does not return value. It is generating new file adding
  language code before .vtt
}
\description{
Translation function for vtt files. Function will be able to send your vtt
file for translation in Google Translate API and return proper strucuture of
the file back. Note: Google Tranlation with API is paid service, however
300USD is given for free for 12 month.
}
\details{
To do: use http://r-pkgs.had.co.nz/tests.html to write test for this package
}
\examples{

## Not run:
# send one vtt file for translation
translateVTT(fileName = "L0.vtt",
             sourceLang = "en",
             destLang = "es",
             apikey = google.api.key)

## send multiple files for translations

# create list of files to translate
filesToTranslate <-list.files("~Files/Downloads/", pattern="*.vtt", full.names=TRUE)

# make a list of languages
languages <- c("fr", "tr", "it", "id", "pt", "es", "ms", "de")

# for loop (note translation takes a while 100 translations ~ 1 hour)
for (FILE in filesToTranslate) {
 # for loop for languages
  for (LANG in languages) {
   # translation
   translateVTT(fileName = FILE, sourceLang = "en", destLang = LANG, apikey = api_key)
  }
}

## End(Not run)

}

