% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateVTT.R
\name{translateVTT}
\alias{translateVTT}
\title{Translation function for vtt files.}
\usage{
translateVTT(fileName, sourceLang = "en", destLang, apikey)
}
\arguments{
\item{fileName}{String with a path to the file you want to translate.
Only one file name is accepted at the time}

\item{sourceLang}{Source language code, 'en' is default value.}

\item{destLang}{Destination language for translator function. Only one at the
time}

\item{apikey}{Google Translate API key}
}
\value{
Function does not return value. It is generating new file adding
  language code before .vtt
}
\description{
Function will be able to send vtt file for translation
in Google Translate API and return proper strucuture of
the file back.
}
\details{
Google Tranlation with API is paid service, however
300USD is given for free for 12 month, check clould.google.com
}
\examples{

\donttest{

library(openssl)
library(readr)
library(translateVTT)
library(translateR)

# retrieve filename
file_name <- system.file("extdata", "L0.vtt", package = "translateVTT")

path_ssh <- normalizePath(tempdir(),winslash = "/")
#path_ssh <- "C:/Users/fxtrams/.ssh"

path_private_key <- file.path(path_ssh, "id_api")

# decrypt the key
out <- read_rds(file.path(path_ssh, "api_key.enc.rds"))
google.api.key <- decrypt_envelope(out$data,
                                   out$iv,
                                   out$session,
                                   path_private_key,
                                   password = "") \%>\%
                  unserialize()

# send one vtt file for translation
translateVTT(fileName = file_name,
             sourceLang = "en",
             destLang = "es",
             apikey = google.api.key)

## send multiple files for translations

# create list of files to translate
filesToTranslate <-list.files("~Files/Downloads/", pattern="*.vtt", full.names=TRUE)

# make a list of languages
languages <- c("fr", "tr", "it", "id", "pt", "es", "ms", "de")

# for loop (note translation takes a while 100 translations ~ 1 hour)
for (FILE in filesToTranslate) {
 # for loop for languages
  for (LANG in languages) {
   # translation
   translateVTT(fileName = FILE, sourceLang = "en", destLang = LANG, apikey = api_key)
  }
}

}

}
\author{
(C) 2019 Vladimir Zhbanko
}
